+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ### API
 
+- Add function `ft_erase_range` to erase range of the cells.
 - Add functions to check if table is empty to the API.
 - `ft_ln` returns status of operation.
 - Add new table property `adding_strategy` (2 strategies available - replace(default) and insert).